requirements.txt
networkx>=2.8.4
scikit-learn>=1.0.2
numpy>=1.21.0
torch>=1.9.0
transformers>=4.11.0
psutil>=5.8.0
pyyaml 
PyPDF2 
python-docx 
neo4j 
networkx 
scikit-learn 
numpy 
torch 
transformers 
psutil 
python-louvain